webportal/backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
cryptography==44.0.0
PyJWT==2.10.1
Werkzeug==3.1.3
reportlab==4.4.0
